+++ src/ui/about.rs
@@ -91,12 +91,10 @@ impl SimpleComponent for AboutDialog {
 
             set_release_notes_version: &APP_VERSION,
             set_release_notes: &[
-                "<p>Fixed</p>",
+                "<p>Changed</p>",
 
                 "<ul>",
-                    "<li>Fixed size of images in the \"Appearance\" preferences section (most noticeable on flatpak)</li>",
-                    "<li>Updated default window size in classic appearance option to match the other launchers</li>",
-                    "<li>Fixed the package name for libwebp-utils on Arch Linux</li>",
+                    "<li>Changed the download and update method to anticipate future migration in the API</li>",
                 "</ul>"
             ].join("\n")
         }